Move single function into runner.go
